Moved get_commands_from_dir function to dump.py
