+++ mysql/toolkit/script/dump.py
@@ -1,8 +1,10 @@
 # Dump SQL commands to files
 import os
+import shutil
+from dirutility import ZipBackup
+from looptools import Timer
 from time import time
 from datetime import datetime
-from looptools import Timer
 
 # Conditional import of multiprocessing module
 # Replace with global import
@@ -79,3 +81,23 @@ def write_text(tup):
     command = _command.strip()
     with open(txt_file, 'w') as txt:
         txt.writelines(command)
+
+
+def get_commands_from_dir(directory, zip_backup=True):
+    """Traverse a directory and read contained SQL files."""
+    # Get SQL script file paths
+    failed_scripts = sorted([os.path.join(directory, fn) for fn in os.listdir(directory) if fn.endswith('.sql')])
+
+    # Read each failed SQL file and append contents to a list
+    print('\tReading SQL scripts from files')
+    commands = []
+    for sql_file in failed_scripts:
+        with open(sql_file, 'r') as txt:
+            sql_command = txt.read()
+        commands.append(sql_command)
+
+    # Remove most recent failures folder after reading
+    if zip_backup:
+        ZipBackup(directory).backup()
+        shutil.rmtree(directory)
+    return commands
